Admin GUI tabulates, Triggers Mostly Implemented
The admin GUI now allows tabulation of the total responses given in the
course of the experiment.

In addition, the responses are now written to a file which adds
'RESPONSES' to whatever is selected for the EEG file.  A future version
may try to reconcile the two streams of data, but for now this is
sufficient.

The automated TMS should now interface with the EEG recording loop,
which is where the most logical place for the TMS stim output is.