Added Outport Signal to Stim, Adjusted  Subject Button Height
The automated TMS setting now trips a flag that causes TMS stims while
recording data.  In addition, the buttons in the post-stim interface
buttons for subject response have been lowered in screen position to
require less of a stretch to reach them.

In addition, the mouse has been rendered invisible on the game screen.