Timestamps, Timed Cursor visibility + reposition
The subject interface now prints the timestamps of the TMS response to
the CSV that contains the TMS responses.

The cursor is now set invisible on the GUI screen at all times except
when the subject is able to respond.  Then, it is set visible and placed
in the center of the screen for easy visibility and access.